Move the services into their own folders. Adjust namespaces.
Remove empty resetter tests file
