+++ src/GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp/Services/RemoveSearchableDocument/RemovePublicationSearchableDocumentsRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services;
+﻿namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services.RemoveSearchableDocument;
 
 public record RemovePublicationSearchableDocumentsRequest
 {
