+++ src/GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp/Services/RemoveSearchableDocument/RemovePublicationSearchableDocumentsResponse.cs
@@ -1,3 +1,3 @@
-﻿namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services;
+﻿namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services.RemoveSearchableDocument;
 
 public record RemovePublicationSearchableDocumentsResponse(IReadOnlyDictionary<Guid, bool> ReleaseIdToDeletionResult);
