+++ src/GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp/Services/RemoveSearchableDocument/RemoveSearchableDocumentRequest.cs
@@ -1,0 +1,6 @@
+﻿namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services.RemoveSearchableDocument;
+
+public record RemoveSearchableDocumentRequest
+{
+    public required Guid ReleaseId { get; init; }
+}
