+++ src/GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp/Services/RemoveSearchableDocument/RemoveSearchableDocumentResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services.RemoveSearchableDocument;
+
+public record RemoveSearchableDocumentResponse(bool Success);
