--- src/GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp/Services/RemoveSearchableDocumentRequest.cs
@@ -1,6 +1,0 @@
-﻿namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services;
-
-public record RemoveSearchableDocumentRequest
-{
-    public required Guid ReleaseId { get; init; }
-}
