--- src/GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp/Services/RemoveSearchableDocumentResponse.cs
@@ -1,3 +1,0 @@
-﻿namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services;
-
-public record RemoveSearchableDocumentResponse(bool Success);
